+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 * 【core   】     修复UrlBuilder地址参数问题（issue#I1UWCA@Gitee）
 * 【core   】     修复StrUtil.toSymbolCase转换问题（issue#1075@Github）
 * 【log    】     修复打印null对象显示{msg}异常问题（issue#1084@Github）
+* 【extra  】     修复ServletUtil.getReader中未关闭的问题
+* 【extra  】     修复QrCodeUtil在新版本zxing报错问题（issue#1088@Github）
 
 -------------------------------------------------------------------------------------------------------------